refactor: clean root surface, dead utils, and orphaned config
Remove dead types, collapse formatters barrel, delete streaming duplicate,
clean navigation, and remove orphaned env config and utilities left over
from prior feature deletions.

Changes:
- Remove 9 unused New* insert types and UserProgress from types/db.ts
- Collapse formatters/ barrel into formatters.ts, delete directory
- Delete streaming/parse-event.ts, inline parser in useRetryGeneration hook
- Remove highlight field and commented-out dropdown from navigation.ts
- Remove renderPlaceholderContentAsHTML from placeholder-content.ts
- Delete utils/url.ts and utils/sanitize.ts (zero imports)
- Remove weeklyHoursRangeLabel from mappers/learningPlans.ts
- Remove orphaned googleAiEnv, curationWeightsEnv, curationEnvSource from env.ts
- Remove dead stripe const and resetMonthlyUsage no-op from stripe/
- Delete 3 related test files
