--- src/lib/utils/url.ts
@@ -1,23 +1,0 @@
-/**
- * Return the hostname for a URL with a leading `www.` removed.
- */
-export function extractDomain(url: string): string | null {
-  try {
-    const parsed = new URL(url);
-    return parsed.hostname.replace(/^www\./, '');
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Checks whether a string is a valid HTTP or HTTPS URL.
- */
-export function isValidHttpUrl(raw: string): boolean {
-  try {
-    const parsed = new URL(raw);
-    return parsed.protocol === 'http:' || parsed.protocol === 'https:';
-  } catch {
-    return false;
-  }
-}
