+++ tests/unit/utils/navigation.spec.ts
@@ -53,11 +53,11 @@ describe('Navigation', () => {
       expect(authenticatedNavItems.length).toBe(4);
     });
 
-    it('should not have highlight flag on any items', () => {
-      const highlightedItems = authenticatedNavItems.filter(
-        (item) => item.highlight
-      );
-      expect(highlightedItems.length).toBe(0);
+    it('should not have any unexpected properties on items', () => {
+      for (const item of authenticatedNavItems) {
+        expect(item).toHaveProperty('label');
+        expect(item).toHaveProperty('href');
+      }
     });
 
     it('should have dropdown only on Analytics item', () => {
@@ -130,11 +130,11 @@ describe('Navigation', () => {
       expect(settingsItem).toBeUndefined();
     });
 
-    it('should not have highlight flag on any items', () => {
-      const highlightedItems = unauthenticatedNavItems.filter(
-        (item) => item.highlight
-      );
-      expect(highlightedItems.length).toBe(0);
+    it('should not have any unexpected properties on items', () => {
+      for (const item of unauthenticatedNavItems) {
+        expect(item).toHaveProperty('label');
+        expect(item).toHaveProperty('href');
+      }
     });
 
     it('should not have dropdown on any items', () => {
